Renamed sample_evn file and fixed logging error bug
