+++ loggings/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class AppLogsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'logging'
+    name = 'loggings'
