+++ logs/app.log.json
@@ -16,3 +16,8 @@
 {"message": "/mnt/c/Users/user/fas_professional_lms/fas_lms_backend/courses/admin.py changed, reloading.", "timestamp": "2024-07-19T06:08:58.243892+00:00", "level": "INFO", "context": {"taskName": null, "name": "django.utils.autoreload", "filename": "autoreload.py", "funcName": "trigger_reload", "msecs": 240.0}}
 {"message": "Watching for file changes with StatReloader", "timestamp": "2024-07-19T06:09:56.958938+00:00", "level": "INFO", "context": {"taskName": null, "name": "django.utils.autoreload", "filename": "autoreload.py", "funcName": "run_with_reloader", "msecs": 955.0}}
 {"message": "Watching for file changes with StatReloader", "timestamp": "2024-07-19T08:04:41.869604+00:00", "level": "INFO", "context": {"taskName": null, "name": "django.utils.autoreload", "filename": "autoreload.py", "funcName": "run_with_reloader", "msecs": 867.0}}
+{"message": "Watching for file changes with StatReloader", "timestamp": "2024-07-20T07:12:31.491950+00:00", "level": "INFO", "context": {"taskName": null, "name": "django.utils.autoreload", "filename": "autoreload.py", "funcName": "run_with_reloader", "msecs": 490.0}}
+{"message": "Watching for file changes with StatReloader", "timestamp": "2024-07-20T07:14:21.432241+00:00", "level": "INFO", "context": {"taskName": null, "name": "django.utils.autoreload", "filename": "autoreload.py", "funcName": "run_with_reloader", "msecs": 430.0}}
+{"message": "Watching for file changes with StatReloader", "timestamp": "2024-07-20T07:15:17.021315+00:00", "level": "INFO", "context": {"taskName": null, "name": "django.utils.autoreload", "filename": "autoreload.py", "funcName": "run_with_reloader", "msecs": 20.0}}
+{"message": "Watching for file changes with StatReloader", "timestamp": "2024-07-20T07:17:45.288144+00:00", "level": "INFO", "context": {"taskName": null, "name": "django.utils.autoreload", "filename": "autoreload.py", "funcName": "run_with_reloader", "msecs": 287.0}}
+{"message": "Hello from log", "timestamp": "2024-07-20T07:18:13.254540+00:00", "level": "INFO", "context": {"taskName": null, "random": 76, "name": "loggings.views", "filename": "views.py", "funcName": "index", "msecs": 252.0}}
