+++ project/settings.py
@@ -12,7 +12,7 @@ https://docs.djangoproject.com/en/5.0/ref/settings/
 
 import os
 from pathlib import Path
-from logging.logger import CustomisedJSONFormatter
+from loggings.logger import CustomisedJSONFormatter
 
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
@@ -74,7 +74,7 @@ INSTALLED_APPS = [
     'connections',
     'companies',
     'certifications',
-    # 'logging',
+    # 'loggings',
     
 ]
